Factor out `ForgeEventTrace` & `TxSubmissionTrace` to their own modules
