+++ cardano-node/src/Cardano/Tracing/ToObjectOrphans/TxSubmissionTrace.hs
@@ -1,0 +1,59 @@
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
+module Cardano.Tracing.ToObjectOrphans.TxSubmissionTrace () where
+
+import           Cardano.Prelude
+
+import           Data.Aeson hiding (Error)
+
+import           Cardano.BM.Data.Tracer (mkObject, trStructured)
+import           Cardano.BM.Tracing ( DefinePrivacyAnnotation
+                                    , DefineSeverity(..), ToObject(..)
+                                    , Severity(..), Transformable(..)
+                                    )
+import           Ouroboros.Consensus.Mempool.API (GenTx, GenTxId)
+import           Ouroboros.Consensus.TxSubmission (TraceLocalTxSubmissionServerEvent(..))
+import           Ouroboros.Network.TxSubmission.Inbound (TraceTxSubmissionInbound)
+import           Ouroboros.Network.TxSubmission.Outbound (TraceTxSubmissionOutbound)
+
+instance DefinePrivacyAnnotation (TraceTxSubmissionInbound
+                                  (GenTxId blk) (GenTx blk))
+instance DefineSeverity (TraceTxSubmissionInbound
+                         (GenTxId blk) (GenTx blk)) where
+  defineSeverity _ = Info
+
+instance DefinePrivacyAnnotation (TraceTxSubmissionOutbound
+                                  (GenTxId blk) (GenTx blk))
+instance DefineSeverity (TraceTxSubmissionOutbound
+                         (GenTxId blk) (GenTx blk)) where
+  defineSeverity _ = Info
+
+instance DefinePrivacyAnnotation (TraceLocalTxSubmissionServerEvent blk)
+instance DefineSeverity (TraceLocalTxSubmissionServerEvent blk) where
+  defineSeverity _ = Info
+
+instance Transformable Text IO (TraceTxSubmissionInbound
+                                (GenTxId blk) (GenTx blk)) where
+  trTransformer _ verb tr = trStructured verb tr
+
+instance Transformable Text IO (TraceTxSubmissionOutbound
+                                (GenTxId blk) (GenTx blk)) where
+  trTransformer _ verb tr = trStructured verb tr
+
+instance Transformable Text IO (TraceLocalTxSubmissionServerEvent blk) where
+  trTransformer _ verb tr = trStructured verb tr
+
+instance ToObject (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)) where
+  toObject _verb _ =
+    mkObject [ "kind" .= String "TraceTxSubmissionInbound" ]
+
+instance ToObject (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)) where
+  toObject _verb _ =
+    mkObject [ "kind" .= String "TraceTxSubmissionOutbound" ]
+
+instance ToObject (TraceLocalTxSubmissionServerEvent blk) where
+  toObject _verb _ =
+    mkObject [ "kind" .= String "TraceLocalTxSubmissionServerEvent" ]
