+++ cardano-node/src/Cardano/Tracing/Tracers.hs
@@ -71,9 +71,11 @@ import           Cardano.Tracing.ToObjectOrphans.ChainDBTrace ()
 import           Cardano.Tracing.ToObjectOrphans.ChainSyncEventTrace ()
 import           Cardano.Tracing.ToObjectOrphans.DnsTrace ()
 import           Cardano.Tracing.ToObjectOrphans.ErrorPolicyTrace ()
+import           Cardano.Tracing.ToObjectOrphans.ForgeEventTrace ()
 import           Cardano.Tracing.ToObjectOrphans.LabelPeerTrace ()
 import           Cardano.Tracing.ToObjectOrphans.MuxTrace ()
 import           Cardano.Tracing.ToObjectOrphans.SubscriptionTrace ()
+import           Cardano.Tracing.ToObjectOrphans.TxSubmissionTrace ()
 import           Cardano.Tracing.ToObjectOrphans
 import           Cardano.Tracing.MicroBenchmarking
